+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Add pagination to `chat` by integrating the use of the `listing` component - [#340](https://github.com/ripe-tech/ripe-robin-revamp/issues/340)
 * Add and adapt to newest chat message placeholder refresh component - [#340](https://github.com/ripe-tech/ripe-robin-revamp/issues/340)
+* Add chat attachment images aggregation according to defined `aggregationThreshold` - [#332](https://github.com/ripe-tech/ripe-robin-revamp/issues/332)
 
 ### Changed